src/user-manager/backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-jose[cryptography] # Libreria per JWT
passlib[bcrypt] # Libreria per l'hashing delle password